Устранение дублирования кода
